\name{plus}
\alias{plus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Add two numbers together }
\description{
  Simple function for testing RtTests package
}
\usage{
plus(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ one argument }
  \item{y}{ the other }
}
\details{
  returns the result of \code{x + y}
}
\value{
  A numeric vector.
}
\author{ Tony Plate }
\examples{
1 + 2
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }
